moves getters into concrete classes.
the static code analyzer was nagging, let's appease it.
