+++ src/Entity/Decklist.php
@@ -245,6 +245,14 @@ class Decklist extends BaseDeck implements JsonSerializable
     }
 
     /**
+     * @inheritdoc
+     */
+    public function getId()
+    {
+        return $this->id;
+    }
+
+    /**
      * Set name
      *
      * @param string $name
@@ -256,6 +264,14 @@ class Decklist extends BaseDeck implements JsonSerializable
         $this->name = $name;
 
         return $this;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getName()
+    {
+        return $this->name;
     }
 
     /**
@@ -297,6 +313,14 @@ class Decklist extends BaseDeck implements JsonSerializable
     }
 
     /**
+     * @inheritdoc
+     */
+    public function getDateCreation()
+    {
+        return $this->dateCreation;
+    }
+
+    /**
      * Set dateUpdate
      *
      * @param DateTime $dateUpdate
@@ -311,6 +335,14 @@ class Decklist extends BaseDeck implements JsonSerializable
     }
 
     /**
+     * @inheritdoc
+     */
+    public function getDateUpdate()
+    {
+        return $this->dateUpdate;
+    }
+
+    /**
      * Set descriptionMd
      *
      * @param string $descriptionMd
@@ -322,6 +354,14 @@ class Decklist extends BaseDeck implements JsonSerializable
         $this->descriptionMd = $descriptionMd;
 
         return $this;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getDescriptionMd()
+    {
+        return $this->descriptionMd;
     }
 
     /**
